Develop

+++ vampireclonevivorsgame/Assets/Scripts/Enemy N1.cs
@@ -9,7 +9,7 @@ public class EnemyN1 : MonoBehaviour
 {
     private float velocidadE = 2.5f;
     private GameObject player;
-    private int vidaEn1 = 3;
+    private float vidaEn1 = 3;
     public static int killslvl1 = 0;
     private float dañoP = 1f;
     // Start is called before the first frame update
